move album CRUD hooks to features folder
