+++ src/app/albums/AlbumsPage.tsx
@@ -5,7 +5,7 @@ import AlbumCard from "../../components/AlbumCard";
 import { openEditAlbumModal } from "@/components/EditAlbumModal";
 import CardGallery from "@/components/CardGallery";
 import { Album } from "@/types/albums/albumTypes";
-import useAlbums from "@/hooks/albums/useAlbums";
+import useAlbums from "@/features/albums/albumManagement/useAlbums";
 import ErrorPage from "../error";
 import LoadingPage from "../loading";
 import {
